STYLE: removed a warning
git-svn-id: https://svn.rwth-aachen.de/repos/vrgroup-svn/projects/Vista/trunk@29743 33c38b33-09d2-4561-8d4d-1bfa2f58f885

+++ VistaCoreLibs/VistaDeviceDriversBase/VistaDeviceDriver.cpp
@@ -130,7 +130,7 @@ void IVistaDeviceDriver::SwapSensorMeasures(std::vector<int> *liIds )
 	{
 		if( (*it)->SwapMeasures() && liIds )
 		{
-			(*liIds).push_back( it - m_vecSensors.begin() );
+			(*liIds).push_back( (int)( it - m_vecSensors.begin() ) );
 		}
 	}
 }
